refactor: log errors and exceptions
